added user validation for update developer and managerapi, create developer and manger api
