+++ log/error.log
@@ -66,3 +66,14 @@ Jan-22-2023 15:38:07 [error] 500 - Cannot create property 'updatedAt' on string 
 Jan-22-2023 15:43:23 [error] 500 - Cannot create property 'updatedAt' on string 'completed' 
 Jan-22-2023 15:45:17 [error] 500 - Cannot create property 'updatedAt' on string 'completed' 
 Jan-22-2023 15:49:00 [error] 404 - Project is already completed 
+Jan-23-2023 10:22:59 [error] 412 - email is required 
+Jan-23-2023 10:23:54 [error] 500 - Joi.string(...).email(...).lowercase(...).unique is not a function 
+Jan-23-2023 10:25:12 [error] 412 - first_nam is not allowed 
+Jan-23-2023 10:25:23 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:26:52 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:27:22 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:27:24 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:27:58 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:30:06 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:32:15 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:32:55 [error] 412 - first_na is not allowed
