+++ log/general.log
@@ -202,3 +202,30 @@ Jan-22-2023 15:45:17 [error] 500 - Cannot create property 'updatedAt' on string 
 Jan-22-2023 15:46:31 [info] http://localhost:5000/api/project/updateProject/63cd0ae1ecb8b03d24664b36 - PUT METHOD  
 Jan-22-2023 15:48:50 [info] http://localhost:5000/api/project/updateProject/63cd0ae1ecb8b03d24664b36 - PUT METHOD  
 Jan-22-2023 15:49:00 [error] 404 - Project is already completed 
+Jan-23-2023 10:22:59 [info] http://localhost:5000/api/user/login - POST METHOD  
+Jan-23-2023 10:22:59 [error] 412 - email is required 
+Jan-23-2023 10:23:16 [info] http://localhost:5000/api/user/login - POST METHOD  
+Jan-23-2023 10:23:54 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:23:54 [error] 500 - Joi.string(...).email(...).lowercase(...).unique is not a function 
+Jan-23-2023 10:25:12 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:25:12 [error] 412 - first_nam is not allowed 
+Jan-23-2023 10:25:23 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:25:23 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:26:05 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:26:52 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:26:52 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:27:22 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:27:22 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:27:23 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:27:24 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:27:50 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:27:58 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:28:42 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:30:06 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:30:06 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:31:29 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:32:15 [error] 500 - Cannot read properties of undefined (reading 'length') 
+Jan-23-2023 10:32:55 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:32:55 [error] 412 - first_na is not allowed 
+Jan-23-2023 10:33:33 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD  
+Jan-23-2023 10:34:30 [info] http://localhost:5000/api/user/updateProfile - PUT METHOD
