+++ middleware/joeValidator.js
@@ -1,16 +1,11 @@
 const appError = require("../helper/errorHandlers/errorHandler");
-const {ErrorMessage}=require('../helper/message');
-const {ErrorCode}=require('../helper/statusCode');
 
 
 module.exports.validationMiddleware = (validator)=> (req,res,next)=>{
     try {
         const {error}=validator(req.body)
-        
         if (error) {
-            throw new appError( error.details?.[0]?.message.replace(/"/g, ""), 412)
-            
-                
+            throw new appError( error.details?.[0]?.message.replace(/"/g, ""), 412)     
         }
         next()
     } catch (error) {
