+++ routes/userRoute/user.route.js
@@ -3,9 +3,9 @@ const { upload } = require("../../utils/aws/aws");
 const auth = require("../../middleware/auth");
 const userController = require("../../controllers/user.controller");
 const {validationMiddleware}=require("../../middleware/joeValidator")
-const {logIn} = require("../../validator/admin.validator")
+const {logIn,createDeveloper,updateData,} = require("../../validator/user.validator")
 
-router.post("/addDeveloper", auth.verifyToken, userController.addDeveloper);
+router.post("/addDeveloper", auth.verifyToken,validationMiddleware(createDeveloper), userController.addDeveloper);
 
 router.post('/login',validationMiddleware(logIn), userController.login);
 
@@ -16,6 +16,7 @@ router.put(
   "/updateProfile",
   auth.verifyToken,
   upload.array("profile_image"),
+  validationMiddleware(updateData),
   userController.updateProfile
 ); //update api for manager and developer
